fix(verification): fix list verification (#51)
Some models can output a list of instances. Sagemaker checks for the length and rejects it if it does not match the input length (1!=N).
In that case we can wrap the output list in another. But then the verification of the outputs fails. This Commit changes that and allows wrapping.