feat: resolve data URIs in Base AI (#560)
* add baseAI version to metadata

* add api functions

* fix tests

* make function simpler

* add data resolver

* fix tests, combine pytest logic, output coverage by default

* extend test

* Add lcov.info to .gitignore

* recover pytest.ini

* fix cassette

* fix bug and add uploader

* Refactor prediction result upload logic

* fix: remove the historic prediction instance logic.

# BREAKING CHANGE

Before each prediction call could return multiple instances. This was born out of the requirements of the time but this was never covered by any schema and complicated the logic in the SDK and backend and also was not intuitive for model developers. Our models currently dont use it anymore and always return a single instance.
This change will add a warning if a list is returned and will pick the first list element as the output.

* add try/catch blocks around download and upload logic to allow backwards compatible deployment

* fix sonar issues

* fix: set base ai version metadata in AI init

* adresss review comments

* feat: add downloading of ref based payloads

* only download when having dict as input

* correctly pass keys in input for backwards compatibility, refactor classes

* fix sonar issues

* fix: put score in metadata and only upload main prediction payload to data storage

* fix test

* feat: add convenience method for file upload